# Add support for CAST(DECIMAL as VARCHAR)