--- a/velox/benchmarks/basic/CastBenchmark.cpp
+++ b/velox/benchmarks/basic/CastBenchmark.cpp
@@ -27,17 +27,31 @@ int main(int argc, char** argv) {
   folly::Init init(&argc, &argv);
 
   ExpressionBenchmarkBuilder benchmarkBuilder;
-
+  const vector_size_t vectorSize = 1000;
   auto vectorMaker = benchmarkBuilder.vectorMaker();
   auto invalidInput = vectorMaker.flatVector<facebook::velox::StringView>({""});
   auto validInput = vectorMaker.flatVector<facebook::velox::StringView>({""});
   auto nanInput = vectorMaker.flatVector<facebook::velox::StringView>({""});
+  auto decimalInput = vectorMaker.flatVector<int64_t>(
+      vectorSize, [&](auto j) { return 12345 * j; }, nullptr, DECIMAL(9, 2));
+  auto shortDecimalInput = vectorMaker.flatVector<int64_t>(
+      vectorSize,
+      [&](auto j) { return 123456789 * j; },
+      nullptr,
+      DECIMAL(18, 6));
+  auto longDecimalInput = vectorMaker.flatVector<int128_t>(
+      vectorSize,
+      [&](auto j) {
+        return facebook::velox::HugeInt::build(12345 * j, 56789 * j + 12345);
+      },
+      nullptr,
+      DECIMAL(38, 16));
 
-  invalidInput->resize(1000);
-  validInput->resize(1000);
-  nanInput->resize(1000);
+  invalidInput->resize(vectorSize);
+  validInput->resize(vectorSize);
+  nanInput->resize(vectorSize);
 
-  for (int i = 0; i < 1000; i++) {
+  for (int i = 0; i < vectorSize; i++) {
     nanInput->set(i, "$"_sv);
     invalidInput->set(i, StringView::makeInline(std::string("")));
     validInput->set(i, StringView::makeInline(std::to_string(i)));
@@ -45,9 +59,20 @@ int main(int argc, char** argv) {
 
   benchmarkBuilder
       .addBenchmarkSet(
-          "cast_int",
+          "cast",
           vectorMaker.rowVector(
-              {"valid", "empty", "nan"}, {validInput, invalidInput, nanInput}))
+              {"valid",
+               "empty",
+               "nan",
+               "decimal",
+               "short_decimal",
+               "long_decimal"},
+              {validInput,
+               invalidInput,
+               nanInput,
+               decimalInput,
+               shortDecimalInput,
+               longDecimalInput}))
       .addExpression("try_cast_invalid_empty_input", "try_cast (empty as int) ")
       .addExpression(
           "tryexpr_cast_invalid_empty_input", "try (cast (empty as int))")
@@ -56,6 +81,10 @@ int main(int argc, char** argv) {
       .addExpression("try_cast_valid", "try_cast (valid as int)")
       .addExpression("tryexpr_cast_valid", "try (cast (valid as int))")
       .addExpression("cast_valid", "cast(valid as int)")
+      .addExpression(
+          "cast_decimal_to_inline_string", "cast (decimal as varchar)")
+      .addExpression("cast_short_decimal", "cast (short_decimal as varchar)")
+      .addExpression("cast_long_decimal", "cast (long_decimal as varchar)")
       .withIterations(100)
       .disableTesting();