--- a/velox/expression/CastExpr.h
+++ b/velox/expression/CastExpr.h
@@ -248,21 +248,12 @@ class CastExpr : public SpecialForm {
       const BaseVector& input,
       VectorPtr& result);
 
-  template <typename TOutput>
-  void applyVarCharToDecimalCastKernel(
-      const SelectivityVector& rows,
-      const BaseVector& input,
-      exec::EvalCtx& context,
-      const TypePtr& toType,
-      VectorPtr castResult);
-
   template <typename FromNativeType>
   VectorPtr applyDecimalToVarcharCast(
       const SelectivityVector& rows,
       const BaseVector& input,
       exec::EvalCtx& context,
-      const TypePtr& fromType,
-      const TypePtr& toType);
+      const TypePtr& fromType);
 
   template <TypeKind ToKind>
   void applyCastPrimitivesDispatch(