To improve the performance, instead of creating intermediate strings, the 
raw string buffer is pre-allocated to be written directly. Function `std::to_chars`
is used to convert integers into a character string by successively filling
the range. On buffer allocation, instead of calculating the precise size 
from intermediate strings, we pre-allocate sufficient buffer based on an 
estimation with decimal precision and scale, and set the precise size after all 
strings are written.

An alternative implementation used `DecimalUtil::toString` which produced a lot 
of intermediate strings during conversion. Besides, `DecimalUtil::toString` 
was called for the calculation of string buffer size. The optimized implementation
uses `std::to_chars` to convert integer to string and avoid all intermediate strings.
The string buffer size is estimated with decimal precision and scale. As below
benchmarks show, the final performance is improved 4-5x compared with the
previous one.

Cast from decimal to varchar benchmark | cast##cast_short_decimal | cast##cast_long_decimal
-- | -- | --
previous (DecimalUtil::toString) | 45.43ms | 132.09ms
optimized (std::to_chars) | 9.87ms | 35.00ms