@jinchengchenghh 2023-08-31
Please look at this issue, this function cannot throw exception too. https://github.com/facebookincubator/velox/discussions/6323