@rui-mo 2023-09-01
Just updated this PR, and the main change is to remove offset and update the pointer directly. Benchmark results (similar as last): 
```
cast##cast_short_decimal                                    9.90ms
cast##cast_long_decimal                                    35.16ms
```